requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-dotenv
chromadb
langchain-huggingface
langchain-openai
langchain-text-splitters
sentence-transformers
